+++ projects/ngx-printer/src/lib/print-item-marker.directive.spec.ts
@@ -5,7 +5,7 @@ import { TestBed } from '@angular/core/testing';
 describe('PrintItemMarkerDirective', () => {
   it('should create an instance', () => {
     const service = TestBed.get(NgxPrinterService);
-    const directive = new PrintItemMarkerDirective(null, service);
+    const directive = new PrintItemMarkerDirective(null, null, service);
     expect(directive).toBeTruthy();
   });
 });